RISC-PVM 0.7.1 => move refine functions into refine directors.
