+++ src/risc-pvm/interpreter/host/handlers/refine/peekHandler.ts
@@ -1,10 +1,10 @@
-import { readBytes, writeBytes } from "../../instructions/helpers";
-import { ExitReasonType } from "../../types";
-import { OK, OOB, WHO } from "../consts";
-import { finish } from "../helpers";
-import { ensureInnerMem, innerRead } from "../innerMem/helpers";
-import { InnerMachineState } from "../innerMem/types";
-import { HostCallHandler } from "../types";
+import { readBytes, writeBytes } from "../../../instructions/helpers";
+import { ExitReasonType } from "../../../types";
+import { OK, OOB, WHO } from "../../consts";
+import { finish } from "../../helpers";
+import { ensureInnerMem, innerRead } from "../../innerMem/helpers";
+import { InnerMachineState } from "../../innerMem/types";
+import { HostCallHandler } from "../../types";
 
 // PEEK (ΩK) — selector 9
 export const peekHandler: HostCallHandler = (s, _id, env) => {
