+++ src/risc-pvm/interpreter/host/hostCallHandlers.ts
@@ -8,13 +8,13 @@ import { writeHandler } from "./handlers/general/writeHandler";
 import { infoHandler } from "./handlers/general/infoHandler";
 import { HostEnvInterface } from "./hostEnvInterface";
 import { HostCallHandler } from "./types";
-import { historicalLookupHandler } from "./handlers/historicalLookupHandler";
-import { exportHandler } from "./handlers/exportHandler";
-import { machineHandler } from "./handlers/machineHandler";
-import { peekHandler } from "./handlers/peekHandler";
-import { pokeHandler } from "./handlers/pokeHandler";
-import { invokeHandler } from "./handlers/invokeHandler";
-import { expungeHandler } from "./handlers/expungeHandler";
+import { historicalLookupHandler } from "./handlers/refine/historicalLookupHandler";
+import { exportHandler } from "./handlers/refine/exportHandler";
+import { machineHandler } from "./handlers/refine/machineHandler";
+import { peekHandler } from "./handlers/refine/peekHandler";
+import { pokeHandler } from "./handlers/refine/pokeHandler";
+import { invokeHandler } from "./handlers/refine/invokeHandler";
+import { expungeHandler } from "./handlers/refine/expungeHandler";
 import { blessHandler } from "./handlers/accumulate/blessHandler";
 import { assignHandler } from "./handlers/accumulate/assignHandler";
 import { designateHandler } from "./handlers/accumulate/designateHandler";
@@ -22,7 +22,7 @@ import { checkpointHandler } from "./handlers/accumulate/checkpointHandler";
 import { newHandler } from "./handlers/accumulate/newHandler";
 import { upgradeHandler } from "./handlers/accumulate/upgradeHandler";
 import { transferHandler } from "./handlers/accumulate/transferHandler";
-import { pagesHandler } from "./handlers/pagesHandler";
+import { pagesHandler } from "./handlers/refine/pagesHandler";
 import { ejectHandler } from "./handlers/accumulate/ejectHandler";
 import { queryHandler } from "./handlers/accumulate/queryHandler";
 import { yieldHandler } from "./handlers/accumulate/yieldHandler";
